api/requirements.txt
Flask==2.1.1
kafka-python==2.0.2
Werkzeug==2.2.2